PB-41 Added possibilities fro adding operation

+++ src/Peppa/Contracts/Requests/CreateTransferOperationRequest.cs
@@ -5,9 +5,9 @@ namespace Peppa.Contracts.Requests
     public class CreateTransferOperationRequest
     {
         public DateTime OperationDate { get; set; }
-        public long From { get; set; }
-        public long To { get; set; }
-        public long Amount { get; set; }
+        public int From { get; set; }
+        public int To { get; set; }
+        public decimal Amount { get; set; }
         public string Comment { get; set; }
     }
 }
